Add EimageIsrTask and modify ProcessEimageTask to use it
As part of refactoring ProcessCcdTask the preferred way to handle custom ISR
(e.g. loading e-images as post-ISR images) is to implement a custom ISR task
instead of a custom ProcessCcdTask. In this case the custom task also appears
useful, since a separate binary and defaults are wanted, but it becomes nearly trivial,
and primarily retargets the ISR task with the new EimageIsrTask.