Add mapper for parquet tables for QA
DM-12030 adds parquet table persistance to the pipe_analysis scripts for
the purpose of interactive QA plotting in a Bokeh based system under
development.  Here, we add the necessary Butler mappers so that they can
get the right filenames.  In order to write these tables, you need to have
the fastparquet python library. If it is not installed, the table files
simply will not get written and you will be warned accordingly.